remove pmigration and  use getAllPool instead of getYourPool Api
