+++ src/components/Pools/Stake/SelectPoolDrawer/index.tsx
@@ -3,7 +3,7 @@ import React, { useCallback } from 'react';
 import AutoSizer from 'react-virtualized-auto-sizer';
 import { FixedSizeList } from 'react-window';
 import Drawer from 'src/components/Drawer';
-import { useGetUserLP } from 'src/state/pmigrate/hooks';
+import { useGetUserLP } from 'src/state/pwallet/hooks';
 import PoolRow from './PoolRow';
 import { CurrencyList } from './styled';
