+++ src/hooks/Tokens.ts
@@ -255,7 +255,7 @@ export function useCoinGeckoTokenPrice(coin: Token) {
       try {
         const chain = coin.chainId === 43113 ? CHAINS[ChainId.AVALANCHE] : CHAINS[coin.chainId];
 
-        const url = `${COINGEKO_BASE_URL}simple/token_price/${
+        const url = `${COINGEKO_BASE_URL}/simple/token_price/${
           chain.coingecko_id
         }?contract_addresses=${coin.address.toLowerCase()}&vs_currencies=usd`;
         const response = await fetch(url);
@@ -282,7 +282,7 @@ export function useCoinGeckoTokenPriceChart(coin: Token, days = '7') {
       try {
         const chain = coin.chainId === 43113 ? CHAINS[ChainId.AVALANCHE] : CHAINS[coin.chainId];
 
-        const url = `${COINGEKO_BASE_URL}coins/${
+        const url = `${COINGEKO_BASE_URL}/coins/${
           chain.coingecko_id
         }/contract/${coin.address.toLowerCase()}/market_chart/?vs_currency=usd&days=${days}`;
